Lesson_11/requirements.txt
pytest~=8.4.2
selenium~=4.36.0
allure-pytest~=2.15.0
requests~=2.32.5